Change health to be a 16-bit value Hopefully I didn't miss anything

+++ src/include/entity.h
@@ -4,7 +4,6 @@
 #define _entity_h_INCLUDE
 
 /// @brief Contains minimum required information for an entity.
-/// @note Takes 8 bytes.
 typedef struct {
     // The entity's position in the overworld.
     uint16_t x;
@@ -22,26 +21,26 @@ typedef struct {
 } Entity;
 
 /// @brief Contains all battle-specific entity information.
-/// @note Takes 8 bytes.
 typedef struct {
     // The maximum health of the encounterable. The health of the encounterable cannot go higher than this.
-    uint8_t max_health;
+    uint16_t max_health;
 
     // Health of the encounterable, which for the purposes of simplifying code, includes the player. If it reaches zero, it dies.
-    uint8_t health;
+    uint16_t health;
 
     // The INT stat of the encounterable.
-    // This has a slight effect on damage, and limits what weapons can be used.
+    // This affects how often you get vital crits and misses, and limits what weapons can be used.
     // For zombies this is always zero, since zombies are mostly depicted of not being capable of much of any thought.
     uint8_t intelligence;
 
     // The STR stat of the encounterable.
-    // This has a moderate effect on damage, and limits what weapons can be used.
+    // This affects base damage, and limits what weapons can be used.
     // Often higher for zombies, since their brains wouldn't be there to limit them, like it does for us mere mortals.
     uint8_t strength;
 
     // The AGI stat of the encounterable.
-    // Doesn't have any effect on damage or what weapons can be used, and instead determines turn order. Essentially, higher speeds will let you have a turn more often than your enemy.
+    // This affects turn order and how often you get crits.
+    // For more detail on turn order see lib/game/encounter.c
     // I'm not sure how this is gonna be balanced between the player and the zombies yet.
     uint8_t speed;
